+++ src/app/common/components/tile-base-widget/tile-base-widget.component.ts
@@ -35,6 +35,7 @@ export class TileBaseWidgetComponent {
                 return "failed";
             case "FlowAbortedResult":
                 return "aborted";
+            /* istanbul ignore next */
             default:
                 return "Unknown outcome typename";
         }