+++ src/app/dataset-view/dataset-view-menu/dataset-view-menu.component.spec.ts
@@ -1,9 +1,7 @@
-import { DatasetNavigationInterface } from "../dataset-view.interface";
 import { FormsModule } from "@angular/forms";
 import { ComponentFixture, TestBed } from "@angular/core/testing";
 import { MatMenuModule } from "@angular/material/menu";
 import { DatasetViewMenuComponent } from "./dataset-view-menu.component";
-import { emitClickOnElementByDataTestId } from "src/app/common/base-test.helpers.spec";
 import { BrowserAnimationsModule } from "@angular/platform-browser/animations";
 import { MatButtonToggleModule } from "@angular/material/button-toggle";
 import { AngularSvgIconModule } from "angular-svg-icon";
@@ -17,21 +15,13 @@ import { MatDividerModule } from "@angular/material/divider";
 import { MatTooltipModule } from "@angular/material/tooltip";
 import { Apollo, ApolloModule } from "apollo-angular";
 import { ApolloTestingModule } from "apollo-angular/testing";
+import { DatasetViewTypeEnum } from "../dataset-view.interface";
+import { RouterModule } from "@angular/router";
+import { SharedTestModule } from "src/app/common/shared-test.module";
 
 describe("DatasetViewMenuComponent", () => {
     let component: DatasetViewMenuComponent;
     let fixture: ComponentFixture<DatasetViewMenuComponent>;
-
-    const mockNavigationObject = {
-        navigateToOverview: () => null,
-        navigateToData: () => null,
-        navigateToMetadata: () => null,
-        navigateToHistory: () => null,
-        navigateToLineage: () => null,
-        navigateToDiscussions: () => null,
-        navigateToSettings: () => null,
-        navigateToFlows: () => null,
-    };
 
     beforeEach(async () => {
         await TestBed.configureTestingModule({
@@ -50,6 +40,8 @@ describe("DatasetViewMenuComponent", () => {
                 MatTooltipModule,
                 ApolloModule,
                 ApolloTestingModule,
+                RouterModule,
+                SharedTestModule,
             ],
             declarations: [DatasetViewMenuComponent, DataAccessPanelComponent],
         }).compileComponents();
@@ -57,29 +49,13 @@ describe("DatasetViewMenuComponent", () => {
         fixture = TestBed.createComponent(DatasetViewMenuComponent);
         component = fixture.componentInstance;
         component.datasetBasics = mockDatasetBasicsDerivedFragment;
-        component.datasetNavigation = mockNavigationObject;
         component.datasetPermissions = mockFullPowerDatasetPermissionsFragment;
+        component.datasetViewType = DatasetViewTypeEnum.Metadata;
 
         fixture.detectChanges();
     });
 
     it("should create", () => {
         expect(component).toBeTruthy();
-    });
-
-    Object.keys(mockNavigationObject).forEach((item) => {
-        it(`should check ${item} on click button`, () => {
-            const navigateSpy = spyOn(
-                component.datasetNavigation,
-                item as keyof DatasetNavigationInterface,
-            ).and.callThrough();
-            emitClickOnElementByDataTestId(fixture, `${item}`);
-            fixture.detectChanges();
-            if (item === "navigateToHistory" || item === "navigateToMetadata") {
-                expect(navigateSpy).toHaveBeenCalledWith(1);
-            } else {
-                expect(navigateSpy).toHaveBeenCalledWith();
-            }
-        });
     });
 });