+++ src/app/dataset-view/dataset-view-menu/dataset-view-menu.component.ts
@@ -10,7 +10,7 @@ import {
     ViewChild,
 } from "@angular/core";
 import { MatSidenav } from "@angular/material/sidenav";
-import { DatasetNavigationInterface, DatasetViewTypeEnum } from "../dataset-view.interface";
+import { DatasetViewTypeEnum } from "../dataset-view.interface";
 import { SideNavHelper } from "../../common/sidenav.helper";
 import { isMobileView, promiseWithCatch } from "src/app/common/app.helpers";
 import { DatasetBasicsFragment, DatasetPermissionsFragment } from "src/app/api/kamu.graphql.interface";
@@ -29,9 +29,9 @@ export class DatasetViewMenuComponent implements OnInit, AfterViewInit {
 
     @Input() datasetBasics: DatasetBasicsFragment;
     @Input() datasetPermissions: DatasetPermissionsFragment;
-    @Input() datasetNavigation: DatasetNavigationInterface;
     @Input() datasetViewType: DatasetViewTypeEnum;
     @Input() isMinimizeSearchAdditionalButtons: boolean;
+    public readonly DatasetViewTypeEnum: typeof DatasetViewTypeEnum = DatasetViewTypeEnum;
 
     private sideNavHelper: SideNavHelper;
 
@@ -89,38 +89,8 @@ export class DatasetViewMenuComponent implements OnInit, AfterViewInit {
         return this.datasetPermissionsServices.shouldAllowSettingsTab(this.datasetPermissions);
     }
 
-    public onNavigateToOverview(): void {
-        this.datasetNavigation.navigateToOverview();
-    }
-
-    public onNavigateToData(): void {
-        this.datasetNavigation.navigateToData();
-    }
-
-    public onNavigateToMetadata(): void {
-        this.datasetNavigation.navigateToMetadata(1);
-    }
-
-    public onNavigateToHistory(): void {
-        this.datasetNavigation.navigateToHistory(1);
-    }
-
-    public onNavigateToLineage(): void {
-        this.datasetNavigation.navigateToLineage();
-    }
-
-    public onNavigateToDiscussions(): void {
-        this.datasetNavigation.navigateToDiscussions();
-    }
-
-    public onNavigateToFlows(): void {
-        this.datasetNavigation.navigateToFlows();
-    }
-
-    public onNavigateToSettings(): void {
-        if (this.shouldAllowSettingsTab) {
-            this.datasetNavigation.navigateToSettings();
-        }
+    public get datasetLink(): string {
+        return `/${this.datasetBasics.owner.accountName}/${this.datasetBasics.name}/`;
     }
 
     @HostListener("window:resize")